requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
av==11.0.0
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cookiecutter==2.6.0
ctranslate2==4.2.1
fastapi==0.110.3
faster-whisper==1.0.1
filelock==3.14.0
flatbuffers==24.3.25
flet==0.22.0
flet-core==0.22.0
flet-runtime==0.22.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
packaging==23.2
pathlib==1.0.1
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
text-unidecode==1.3
tokenizers==0.15.2
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
